Extractor: added Rabbitstream
